Convert misc actions to TypeScript
