+++ frontend/src/metabase/dashboard/actions/cards.js
@@ -1,87 +1,9 @@
-import _ from "underscore";
 import { t } from "ttag";
-import { createAction, createThunkAction } from "metabase/lib/redux";
-
-import Questions from "metabase/entities/questions";
 
 import { createCard } from "metabase/lib/card";
 
-import { autoWireParametersToNewCard } from "metabase/dashboard/actions/auto-wire-parameters/actions";
-import { trackCardCreated, trackQuestionReplaced } from "../analytics";
-import { getDashCardById, getDashboardId } from "../selectors";
-import { isVirtualDashCard } from "../utils";
-import {
-  REMOVE_CARD_FROM_DASH,
-  UNDO_REMOVE_CARD_FROM_DASH,
-  setDashCardAttributes,
-} from "./core";
-import { cancelFetchCardData, fetchCardData } from "./data-fetching";
-import { loadMetadataForDashboard } from "./metadata";
+import { trackCardCreated } from "../analytics";
 import { addDashCardToDashboard } from "./cards-typed";
-
-export const MARK_NEW_CARD_SEEN = "metabase/dashboard/MARK_NEW_CARD_SEEN";
-export const markNewCardSeen = createAction(MARK_NEW_CARD_SEEN);
-
-export const replaceCard =
-  ({ dashcardId, nextCardId }) =>
-  async (dispatch, getState) => {
-    const dashboardId = getDashboardId(getState());
-
-    let dashcard = getDashCardById(getState(), dashcardId);
-    if (isVirtualDashCard(dashcard)) {
-      return;
-    }
-
-    await dispatch(Questions.actions.fetch({ id: nextCardId }));
-    const card = Questions.selectors
-      .getObject(getState(), { entityId: nextCardId })
-      .card();
-
-    await dispatch(
-      setDashCardAttributes({
-        id: dashcardId,
-        attributes: {
-          card,
-          card_id: card.id,
-          series: [],
-          parameter_mappings: [],
-          visualization_settings: {},
-        },
-      }),
-    );
-
-    dashcard = getDashCardById(getState(), dashcardId);
-
-    dispatch(fetchCardData(card, dashcard, { reload: true, clearCache: true }));
-    await dispatch(loadMetadataForDashboard([dashcard]));
-    dispatch(autoWireParametersToNewCard({ dashcard_id: dashcardId }));
-
-    trackQuestionReplaced(dashboardId);
-  };
-
-export const removeCardFromDashboard = createThunkAction(
-  REMOVE_CARD_FROM_DASH,
-  ({ dashcardId, cardId }) =>
-    (dispatch, _getState) => {
-      dispatch(cancelFetchCardData(cardId, dashcardId));
-      return { dashcardId };
-    },
-);
-
-export const undoRemoveCardFromDashboard = createThunkAction(
-  UNDO_REMOVE_CARD_FROM_DASH,
-  ({ dashcardId }) =>
-    (dispatch, getState) => {
-      const dashcard = getDashCardById(getState(), dashcardId);
-      const card = dashcard.card;
-
-      if (!isVirtualDashCard(dashcard)) {
-        dispatch(fetchCardData(card, dashcard));
-      }
-
-      return { dashcardId };
-    },
-);
 
 export const addActionToDashboard =
   async ({ dashId, tabId, action, displayType }) =>
